Added to new primitves package: Mesh
