+++ src/main/kotlin/graphics/scenery/Camera.kt
@@ -1,10 +1,10 @@
 package graphics.scenery
 
+import graphics.scenery.primitives.TextBoard
 import graphics.scenery.utils.extensions.minus
 import graphics.scenery.utils.extensions.plus
 import graphics.scenery.utils.extensions.times
 import graphics.scenery.utils.extensions.xyz
-import graphics.scenery.volumes.Volume
 import org.joml.*
 import java.lang.Math
 import java.util.concurrent.atomic.AtomicInteger
