+++ src/main/kotlin/graphics/scenery/CreateCommand.kt
@@ -1,8 +1,6 @@
 package graphics.scenery
 
-import graphics.scenery.Camera
-import graphics.scenery.Mesh
-import graphics.scenery.Scene
+import graphics.scenery.primitives.Mesh
 import graphics.scenery.backends.Renderer
 import graphics.scenery.utils.LazyLogger
 import org.joml.Vector2f
@@ -19,7 +17,8 @@ open class CreateCommand constructor(protected val name: String,
                                            protected val renderer: Renderer,
                                            protected val scene: Scene,
                                            protected val camera: () -> Camera?,
-                                           meshLambda: () -> Mesh) : ClickBehaviour {
+                                           meshLambda: () -> Mesh
+) : ClickBehaviour {
     protected val logger by LazyLogger()
 
     protected val cam: Camera? by CameraDelegate()
