+++ src/main/kotlin/graphics/scenery/Helix.kt
@@ -2,6 +2,7 @@ package graphics.scenery
 
 import graphics.scenery.geometry.Curve
 import graphics.scenery.geometry.Spline
+import graphics.scenery.primitives.Mesh
 import graphics.scenery.proteins.MathLine
 import org.joml.*
