+++ src/main/kotlin/graphics/scenery/backends/opengl/OpenGLRenderer.kt
@@ -11,6 +11,7 @@ import graphics.scenery.*
 import graphics.scenery.backends.*
 import graphics.scenery.geometry.GeometryType
 import graphics.scenery.geometry.HasGeometry
+import graphics.scenery.primitives.Plane
 import graphics.scenery.spirvcrossj.Loader
 import graphics.scenery.spirvcrossj.libspirvcrossj
 import graphics.scenery.textures.Texture
