+++ src/main/kotlin/graphics/scenery/controls/OpenVRHMD.kt
@@ -12,6 +12,7 @@ import graphics.scenery.backends.vulkan.VU
 import graphics.scenery.backends.vulkan.VulkanDevice
 import graphics.scenery.backends.vulkan.VulkanTexture
 import graphics.scenery.backends.vulkan.endCommandBuffer
+import graphics.scenery.primitives.Mesh
 import graphics.scenery.utils.JsonDeserialisers
 import graphics.scenery.utils.LazyLogger
 import kotlinx.coroutines.GlobalScope
