+++ src/main/kotlin/graphics/scenery/effectors/SlicingVolumeEffector.kt
@@ -2,7 +2,7 @@ package graphics.scenery.effectors
 
 import org.joml.Vector3f
 import graphics.scenery.Node
-import graphics.scenery.Plane
+import graphics.scenery.primitives.Plane
 
 /**
  * Slicing volume effector
