+++ src/main/kotlin/graphics/scenery/geometry/CatmullRomSpline.kt
@@ -1,6 +1,5 @@
 package graphics.scenery.geometry
 
-import graphics.scenery.geometry.Spline
 import graphics.scenery.numerics.Random
 import graphics.scenery.utils.extensions.plus
 import graphics.scenery.utils.extensions.times
