+++ src/main/kotlin/graphics/scenery/primitives/Cylinder.kt
@@ -1,5 +1,6 @@
-package graphics.scenery
+package graphics.scenery.primitives
 
+import graphics.scenery.BufferUtils
 import graphics.scenery.geometry.GeometryType
 import org.joml.Vector3f
 import java.util.*
