+++ src/main/kotlin/graphics/scenery/primitives/Plane.kt
@@ -1,8 +1,7 @@
-package graphics.scenery
+package graphics.scenery.primitives
 
+import graphics.scenery.BufferUtils
 import org.joml.Vector3f
-import java.nio.FloatBuffer
-import java.nio.IntBuffer
 
 /**
  * Constructs a plane with the dimensions given in [sizes].
@@ -17,32 +16,40 @@ open class Plane(sizes: Vector3f) : Mesh() {
         val side = 2.0f
         val side2 = side / 2.0f
 
-        vertices = BufferUtils.allocateFloatAndPut(floatArrayOf(
+        vertices = BufferUtils.allocateFloatAndPut(
+            floatArrayOf(
                 // Front
                 -side2, -side2, side2,
                 side2, -side2, side2,
-                side2,  side2, side2,
-                -side2,  side2, side2
-        ))
+                side2, side2, side2,
+                -side2, side2, side2
+            )
+        )
 
-        normals = BufferUtils.allocateFloatAndPut(floatArrayOf(
+        normals = BufferUtils.allocateFloatAndPut(
+            floatArrayOf(
                 // Front
                 0.0f, 0.0f, 1.0f,
                 0.0f, 0.0f, 1.0f,
                 0.0f, 0.0f, 1.0f,
                 0.0f, 0.0f, 1.0f
-        ))
+            )
+        )
 
-        indices = BufferUtils.allocateIntAndPut(intArrayOf(
-                0,1,2,0,2,3
-        ))
+        indices = BufferUtils.allocateIntAndPut(
+            intArrayOf(
+                0, 1, 2, 0, 2, 3
+            )
+        )
 
-        texcoords = BufferUtils.allocateFloatAndPut(floatArrayOf(
+        texcoords = BufferUtils.allocateFloatAndPut(
+            floatArrayOf(
                 0.0f, 0.0f,
                 1.0f, 0.0f,
                 1.0f, 1.0f,
                 0.0f, 1.0f
-        ))
+            )
+        )
 
         boundingBox = generateBoundingBox()
     }
