+++ src/main/kotlin/graphics/scenery/primitives/Skybox.kt
@@ -1,5 +1,7 @@
-package graphics.scenery
+package graphics.scenery.primitives
 
+import graphics.scenery.Box
+import graphics.scenery.ShaderMaterial
 import org.joml.Vector3f
 
 /**
