+++ src/main/kotlin/graphics/scenery/proteins/RibbonDiagram.kt
@@ -7,6 +7,7 @@ import graphics.scenery.geometry.Spline
 import graphics.scenery.geometry.UniformBSpline
 import org.joml.*
 import graphics.scenery.numerics.Random
+import graphics.scenery.primitives.Mesh
 import org.biojava.nbio.structure.Atom
 import org.biojava.nbio.structure.Group
 import org.biojava.nbio.structure.secstruc.SecStrucCalc
