+++ src/test/kotlin/graphics/scenery/tests/examples/advanced/MultiBoxExample.kt
@@ -4,6 +4,7 @@ import org.joml.Vector3f
 import graphics.scenery.*
 import graphics.scenery.backends.Renderer
 import graphics.scenery.numerics.Random
+import graphics.scenery.primitives.Mesh
 import kotlin.concurrent.thread
 
 /**
