+++ src/test/kotlin/graphics/scenery/tests/examples/basic/PointCloudExample.kt
@@ -4,6 +4,7 @@ import org.joml.Vector3f
 import graphics.scenery.*
 import graphics.scenery.backends.Renderer
 import graphics.scenery.numerics.Random
+import graphics.scenery.primitives.PointCloud
 
 /**
  * Simple example to demonstrate the drawing of a 3D point cloud.
