+++ src/test/kotlin/graphics/scenery/tests/examples/cluster/BileExample.kt
@@ -6,6 +6,7 @@ import graphics.scenery.backends.Renderer
 import graphics.scenery.controls.TrackedStereoGlasses
 import graphics.scenery.net.NodePublisher
 import graphics.scenery.net.NodeSubscriber
+import graphics.scenery.primitives.Mesh
 import graphics.scenery.utils.extensions.times
 
 /**
