+++ src/test/kotlin/graphics/scenery/tests/examples/cluster/DemoReelExample.kt
@@ -7,6 +7,7 @@ import graphics.scenery.controls.InputHandler
 import graphics.scenery.controls.TrackedStereoGlasses
 import graphics.scenery.net.NodePublisher
 import graphics.scenery.net.NodeSubscriber
+import graphics.scenery.primitives.Mesh
 import graphics.scenery.volumes.Colormap
 import graphics.scenery.volumes.TransferFunction
 import graphics.scenery.volumes.Volume
