+++ src/test/kotlin/graphics/scenery/tests/examples/volumes/VolumeSamplingExample.kt
@@ -4,6 +4,8 @@ import org.joml.Vector3f
 import graphics.scenery.*
 import graphics.scenery.backends.Renderer
 import graphics.scenery.numerics.Random
+import graphics.scenery.primitives.Cylinder
+import graphics.scenery.primitives.Line
 import graphics.scenery.utils.MaybeIntersects
 import graphics.scenery.utils.RingBuffer
 import graphics.scenery.utils.extensions.minus
