+++ src/test/kotlin/graphics/scenery/tests/unit/CylinderTests.kt
@@ -1,6 +1,6 @@
 package graphics.scenery.tests.unit
 
-import graphics.scenery.Cylinder
+import graphics.scenery.primitives.Cylinder
 import graphics.scenery.Scene
 import graphics.scenery.numerics.Random
 import graphics.scenery.utils.LazyLogger
